requirements.txt
asgiref==3.8.1
azure-core==1.33.0
azure-servicebus==7.14.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
isodate==0.7.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
